requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-filter==2.4.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==20.9
pkg-resources==0.0.0
prompt-toolkit==3.0.18
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
